+++ src-ui/components/SCXTEditor.h
@@ -41,7 +41,6 @@
 #include "sst/jucegui/components/HSlider.h"
 #include "sst/jucegui/components/ToolTip.h"
 #include "sst/jucegui/accessibility/FocusDebugger.h"
-#include "datamodel/adsr_storage.h"
 #include "sst/jucegui/components/WindowPanel.h"
 #include "messaging/client/zone_messages.h"
 #include "infrastructure/rng_gen.h"
